YOLO3_Keras/requirements.txt
absl-py==0.1.11
astor==0.7.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.11.5
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.3
enum34==1.1.6
gast==0.2.0
graphviz==0.8.4
grpcio==1.14.1
h5py==2.7.1
html5lib==0.9999999
imageio==2.1.2
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.4.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.1.0
jupyter-latex-envs==1.4.0
jupyter-nbextensions-configurator==0.4.0
Keras==2.2.0
Keras-Applications==1.0.2
keras-contrib==2.0.8
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.1
moviepy==0.2.3.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.5
olefile==0.44
opencv-python==3.4.0.12
pandas==0.21.1
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.3.0
tensorflow-gpu==1.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.8
